LayoutTests/imported/w3c/web-platform-tests/html/semantics/the-button-element/interest-target/idlharness.tentative-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface Element: member names are unique
PASS Partial interface HTMLAnchorElement: member names are unique
PASS Partial interface HTMLAreaElement: member names are unique
PASS Partial interface SVGAElement: member names are unique
PASS HTMLAnchorElement includes InterestInvokerElement: member names are unique
PASS HTMLAreaElement includes InterestInvokerElement: member names are unique
PASS HTMLButtonElement includes InterestInvokerElement: member names are unique
PASS SVGAElement includes InterestInvokerElement: member names are unique
PASS HTMLElement includes GlobalEventHandlers: member names are unique
PASS HTMLElement includes ElementContentEditable: member names are unique
PASS HTMLElement includes HTMLOrSVGElement: member names are unique
PASS HTMLAnchorElement includes HTMLHyperlinkElementUtils: member names are unique
PASS HTMLAreaElement includes HTMLHyperlinkElementUtils: member names are unique
PASS HTMLButtonElement includes PopoverInvokerElement: member names are unique
PASS Element includes ParentNode: member names are unique
PASS Element includes NonDocumentTypeChildNode: member names are unique
PASS Element includes ChildNode: member names are unique
PASS Element includes Slottable: member names are unique
PASS SVGElement includes GlobalEventHandlers: member names are unique
PASS SVGElement includes SVGElementInstance: member names are unique
PASS SVGElement includes HTMLOrSVGElement: member names are unique
PASS SVGGraphicsElement includes SVGTests: member names are unique
PASS SVGAElement includes SVGURIReference: member names are unique
FAIL InterestEvent interface: existence and properties of interface object assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent interface object length assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent interface object name assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent interface: attribute source assert_own_property: self does not have own property "InterestEvent" expected property "InterestEvent" missing
FAIL InterestEvent must be primary interface of new InterestEvent("interest") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: InterestEvent"
FAIL Stringification of new InterestEvent("interest") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: InterestEvent"
FAIL InterestEvent interface: new InterestEvent("interest") must inherit property "source" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: InterestEvent"
FAIL HTMLAnchorElement interface: attribute interestForElement assert_true: The prototype object must have a property "interestForElement" expected true got false
FAIL HTMLAnchorElement interface: document.createElement("a") must inherit property "interestForElement" with the proper type assert_inherits: property "interestForElement" not found in prototype chain
FAIL HTMLAreaElement interface: attribute interestForElement assert_true: The prototype object must have a property "interestForElement" expected true got false
FAIL HTMLAreaElement interface: document.createElement("area") must inherit property "interestForElement" with the proper type assert_inherits: property "interestForElement" not found in prototype chain
FAIL HTMLButtonElement interface: attribute interestForElement assert_true: The prototype object must have a property "interestForElement" expected true got false
FAIL HTMLButtonElement interface: document.createElement("button") must inherit property "interestForElement" with the proper type assert_inherits: property "interestForElement" not found in prototype chain
FAIL SVGAElement interface: attribute interestForElement assert_true: The prototype object must have a property "interestForElement" expected true got false
FAIL SVGAElement interface: document.createElementNS("http://www.w3.org/2000/svg", "a") must inherit property "interestForElement" with the proper type assert_inherits: property "interestForElement" not found in prototype chain
